Fix race accessing the connection proxy state in finalize()
Accesses to ConnectionProxy.state are unprotected and thus racy. This
leads to a situation where reading 'state' in the finalizer thread
gets an stale value long after it has been set to CLOSED in
ConnectionProxy.invoke in a different thread. This behavior has
been observed in unit tests for products that rely on JDBC lint.

Uses an atomic reference to ensure memory-visibility. The
happens-after relationship is ensured by the fact that the
finalizer only runs after there are no reachable references
to the object. This means, this happens after the invoke
method has been executed.