Fix race accessing the connection proxy state in finalize() (#23)
Fix racy member accesses in finalize()

Accesses to ConnectionProxy.state are unprotected and thus racy. This
leads to a situation where reading 'state' in the finalizer thread
gets an stale value long after it has been set to CLOSED in
ConnectionProxy.invoke in a different thread. This behavior has
been observed in unit tests for products that rely on JDBC lint.

Similar issues are present in BlobProxy, ResultSetProxy and
StatementProxy.

This change will ensure memory visibility in the finalize() method
for the BlobProxy.freed, ConnectionProxy.state,
ResultSetProxy.closed and StatementProxy.state member variables

It uses an atomic references and booleans to ensure memory visibility.
The happens-after relationship is ensured by the fact that the
finalizer only runs after there are no reachable references to the
object. This means, this happens after the invoke method has been
executed.

Additional changes:

* Cleanup: fix indentation
* Return after closing {Connection,Statement,ResultSet} in proxies.
  This avoids invoking close() twice in the underlying objects.

PR #23